requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.4.2
requests>=2.31.0

# Agent system
langchain>=0.0.335
langgraph>=0.0.20
anthropic>=0.8.0
google-generativeai>=0.3.2
openai>=1.2.0
replicate>=0.21.1
faiss-cpu>=1.7.4

# Database (for future integration)
sqlalchemy>=2.0.22
alembic>=1.12.0
psycopg2-binary>=2.9.9

# Utilities
python-multipart>=0.0.6
jinja2>=3.1.2
tenacity>=8.2.3
colorama>=0.4.6
tiktoken>=0.5.1
numpy>=1.26.0
scipy>=1.11.3
pandas>=2.1.1

# Testing
pytest>=7.4.2
httpx>=0.25.0